# Conflicts:
#	sumaris-core/src/main/java/net/sumaris/core/dao/administration/programStrategy/StrategyRepositoryImpl.java
#	sumaris-core/src/main/java/net/sumaris/core/service/data/LandingServiceImpl.java
#	sumaris-core/src/main/resources/net/sumaris/core/db/changelog/hsqldb/db-changelog-1.10.3.xml
#	sumaris-core/src/main/resources/net/sumaris/core/db/changelog/hsqldb/db-changelog.xml

+++ sumaris-core-shared/src/main/java/net/sumaris/core/dao/technical/Page.java
@@ -52,7 +52,7 @@ public class Page implements Serializable {
     private SortDirection sortDirection = SortDirection.ASC;
 
     public Pageable asPageable() {
-        return Pageables.create((int)offset, size, sortBy, sortDirection);
+        return Pageables.create(offset, size, sortBy, sortDirection);
     }
 }
